# Optimized reading the exposure more