--- a/openquake/calculators/tests/multi_risk_test.py
+++ b/openquake/calculators/tests/multi_risk_test.py
@@ -34,7 +34,7 @@ class MultiRiskTestCase(CalculatorTestCase):
 
         # check extract
         md = json.loads(extract(self.calc.datastore, 'exposure_metadata').json)
-        ae(md['names'], ['occupants_night', 'value-number', 'value-structural'])
+        ae(md['names'], ['value-number', 'value-structural', 'occupants_night'])
         ae(md['multi_risk'], ['collapse-structural-ASH_DRY',
                               'collapse-structural-ASH_WET',
                               'loss-structural-ASH_DRY',