… openquake/qa_tests_data/event_based_damage/case_13/expected/d_damages.txt
… openquake/qa_tests_data/multi_risk/case_1/expected/agg_risk.csv