Part of #8761. For the USA with 4+M assets I get a 30% speedup:
```
# before
| calc_62042               | time_sec | memory_mb | counts |
|--------------------------+----------+-----------+--------|
| EventBasedCalculator.run | 108.7    | 1_363     | 1      |
| importing inputs         | 104.7    | 2_476     | 1      |
| reading exposure         | 59.9     | 1_863     | 1      |
| building assetcol        | 24.6     | 611.0     | 1      |
| associating exposure     | 17.2     | 1.28516   | 1      |

# after
| calc_62041               | time_sec | memory_mb | counts |
|--------------------------+----------+-----------+--------|
| EventBasedCalculator.run | 79.9     | 1_386     | 1      |
| importing inputs         | 78.9     | 1_432     | 1      |
| reading exposure         | 64.3     | 1_435     | 1      |
| associating exposure     | 11.4     | 0.0       | 1      |
| building assetcol        | 0.15311  | 0.0       | 1      |
```